+++ duckscript_sdk/src/sdk/std/process/exec/mod_test.rs
@@ -8,7 +8,6 @@ fn common_functions() {
 }
 
 #[test]
-#[test]
 fn run_no_args() {
     test::run_script_and_error(vec![create("")], "out = exec", "out");
 }