+++ duckscript_sdk/src/sdk/std/process/mod.rs
@@ -1,6 +1,7 @@
 mod exec;
 mod exit;
 mod process_id;
+mod watchdog;
 
 use crate::utils::pckg;
 use duckscript::types::command::Commands;
@@ -14,6 +15,7 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(exec::create(&package))?;
     commands.set(exit::create(&package))?;
     commands.set(process_id::create(&package))?;
+    commands.set(watchdog::create(&package))?;
 
     Ok(())
 }